Update patch-steps-lib to 1.1.3
