+++ assets/mods/simplify/lib/patch-steps-lib/AUTHORS
@@ -1,0 +1,3 @@
+20kdc
+ac2pic
+
